+++ R/lollipopPlot2.R
@@ -18,6 +18,9 @@
 #' @param axisTextSize text size for axis labels. Default 1.
 #' @param pointSize size of lollipop heads. Default 1.2
 #' @param domainLabelSize text size for domain labels. Default 1.
+#' @param roundedRect Default FALSE. If `TRUE` domains are drawn with rounded corners. Requires \code{berryFunctions}
+#' @param showDomainLabel Label domains within the plot. Default TRUE. If FALSE domains are annotated in legend.
+#' @param domainBorderCol Default "black". Set to NA to remove.
 #' @param legendTxtSize Default 1.
 #' @param verbose Default TRUE
 #' @importFrom grDevices colors colours
@@ -34,7 +37,7 @@
 #'
 
 lollipopPlot2 = function(m1, m2, gene = NULL, AACol1 = NULL, AACol2 = NULL, m1_name = NULL, m2_name = NULL, m1_label = NULL, m2_label = NULL, refSeqID = NULL, proteinID = NULL, labPosAngle = 0, labPosSize = 0.9,
-                         colors = NULL, alpha = 1, axisTextSize = c(1, 1), pointSize = 1.2, domainLabelSize = 1, legendTxtSize = 1, verbose = TRUE){
+                         colors = NULL, alpha = 1, axisTextSize = c(1, 1), pointSize = 1.2, roundedRect = TRUE, showDomainLabel = TRUE, domainBorderCol = "black", domainLabelSize = 1, legendTxtSize = 1, verbose = TRUE){
 
   if(is.null(gene)){
     stop('Please provide a gene name.')
@@ -96,9 +99,7 @@ lollipopPlot2 = function(m1, m2, gene = NULL, AACol1 = NULL, AACol2 = NULL, m1_n
   }
 
   domains = unique(prot[,Label])
-  domain_cols = c(RColorBrewer::brewer.pal(8, name = "Set2"),
-                  RColorBrewer::brewer.pal(8, name = "Accent"),
-                  RColorBrewer::brewer.pal(12, name = "Set3"))
+  domain_cols = get_domain_cols()
 
   if(length(domains) > length(domain_cols)){
     domain_cols = sample(colours(), size = length(domains), replace = FALSE)
@@ -126,11 +127,11 @@ lollipopPlot2 = function(m1, m2, gene = NULL, AACol1 = NULL, AACol2 = NULL, m1_n
   }
 
   lo = matrix(data = c(1, 1, 2, 2), nrow = 2, byrow = TRUE)
-  graphics::layout(mat = lo, heights = c(4, 1))
+  graphics::layout(mat = lo, heights = c(5, 2))
 
   par(mar = c(2, 2.5, 2, 1))
   plot(0, 0, xlim = c(0, max(prot.len, na.rm = TRUE)+10), ylim = c(-6.5, 6.5), axes = FALSE, pch = NA, xlab = "", ylab = "")
-  rect(xleft = 0, ybottom = -0.8, xright = max(prot.len), ytop = 0.8, col = "gray70", border = "gray70")
+  rect(xleft = 0, ybottom = -0.8, xright = max(prot.len), ytop = 0.8, col = "#95a5a6", border = domainBorderCol)
   # axis(side = 1, at = m1.lp[[2]], labels = m1.lp[[2]], lwd = 1.2, font = 1,
   #      cex.axis = axisTextSize[1], line = -0.4)
   axis(side = 2, at = m1.y.at, labels = m1.y.at.lbl, lwd = 1.2, font = 1, las = 2,
@@ -148,10 +149,29 @@ lollipopPlot2 = function(m1, m2, gene = NULL, AACol1 = NULL, AACol2 = NULL, m1_n
     points(x = m2.lp[[1]][,pos2], y = -m2.lp[[1]][,count2], col = col2, pch = 16, cex = pointSize)
   }
 
-  rect(xleft = prot[,Start], ybottom = -1, xright = prot[,End], ytop = 1, col = domain_cols, border = NA)
+  col_uniq = unique(c(names(col1), names(col2)))
+  col = c(col1, col2)[col_uniq]
+
+  prot[, domainCol := domain_cols[prot[, Label]]]
+  if(roundedRect){
+    if(requireNamespace("berryFunctions", quietly = TRUE)){
+      for(i in 1:nrow(prot)){
+        berryFunctions::roundedRect(xleft = prot[i,Start], ybottom = -1, xright = prot[i,End], ytop = 1, col = prot[i, domainCol], border = domainBorderCol, rounding = 0.08)
+      }
+    }else{
+      #warning("Package berryFunctions needed for roundedRect to work. Please install it and try again.")
+      rect(xleft = prot[,Start], ybottom = -1, xright = prot[,End], ytop = 1, col = prot[,domainCol], border = domainBorderCol)
+    }
+  }else{
+    rect(xleft = prot[,Start], ybottom = -1, xright = prot[,End], ytop = 1, col = prot[,domainCol], border = domainBorderCol)
+  }
+  #rect(xleft = prot[,Start], ybottom = -1, xright = prot[,End], ytop = 1, col = prot[,domainCol], border = NA)
 
   prot$pos = rowMeans(x = prot[,.(Start, End)])
-  text(y = 0, x = prot$pos, labels = prot$Label, font = 3, cex = domainLabelSize)
+  if(showDomainLabel){
+    text(y = 0, x = prot$pos, labels = prot$Label, font = 3, cex = domainLabelSize)
+  }
+
 
   text(x = max(prot.len)+12, y = 0.5, labels = paste0(max(prot.len), " aa"), font = 1, adj = 1, srt = 90)
 
@@ -190,11 +210,28 @@ lollipopPlot2 = function(m1, m2, gene = NULL, AACol1 = NULL, AACol2 = NULL, m1_n
     }
   }
 
-  plot.new()
-  par(mar = c(1, 0, 1, 1))
-  col = col[unique(c(names(col1), names(col2)))]
-  legend("left", legend = names(col), col = col,  bty = "n", border=NA,
-         xpd = TRUE, text.font = 1, pch = 16, xjust = 0, yjust = 0,
-         cex = legendTxtSize, y.intersp = 1.5, x.intersp = 1,
-         pt.cex = 1.2 * legendTxtSize, ncol = ceiling(length(col)/4))
+  par(mar = c(5, 0.5, 0, 0), xpd = TRUE)
+
+  plot(NULL,ylab='',xlab='', xlim=0:1, ylim=0:1, axes = FALSE)
+  lep = legend("topleft", legend = names(col), col = col,  bty = "n", border=NA,
+                 xpd = TRUE, text.font = 1, pch = 16, xjust = 0, yjust = 0,
+                 cex = legendTxtSize, y.intersp = 1.5, x.intersp = 1,
+                 pt.cex = 1.2 * legendTxtSize, ncol = ceiling(length(col)/4))
+
+    x_axp = 0+lep$rect$w
+
+    if(!showDomainLabel){
+      if(length(domain_cols) <= 4){
+        n_col = 1
+      }else{
+        n_col = (length(domain_cols) %/% 4)+1
+      }
+
+      lep = legend(x = x_axp, y = 1, legend = names(domain_cols),
+                   col = domain_cols, border = NA,
+                   ncol= n_col, pch = 15, xpd = TRUE, xjust = 0, bty = "n",
+                   cex = legendTxtSize, title = "Domains",
+                   title.adj = 0, pt.cex = 1.2 * legendTxtSize)
+    }
+
 }